install.sh: Fix ShellCheck warning SC2115
"Use "${var:?}" to ensure this never expands to / ."

https://github.com/koalaman/shellcheck/wiki/SC2115

+++ script/install.sh
@@ -619,7 +619,7 @@ uninstall_compiler() {
         fatal "$1 is not installed in $path"
     fi
     log "Removing $path/$1"
-    rm -rf "$path/$1"
+    rm -rf "${path:?}/$1"
 }
 
 list_compilers() {
